backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
langgraph
langchain
appium-python-client
python-dotenv
